Module Octogram

module Agent : sig ... end

This module introduces the business logic run by an Octogram agent on a remote runner.

module Agent_builtins : sig ... end

This module introduces the remote procedure allowing to manage the lifecycle of a remote agent.

module Agent_name : sig ... end

module Agent_state : sig ... end

module Execution_params : sig ... end

module Global_variables : sig ... end

This modules introduces the table of global variables which can is manipulated by jobs.

module Helpers : sig ... end

module Http_client : sig ... end

module Http_server : sig ... end

Spawns a simple HTTP server (the python 3 module).

module Job : sig ... end

This module introduces the definition of a Octogram job, that is an interaction between the orchestrator and one agent.

module Orchestrator : sig ... end

module Orchestrator_state : sig ... end

module Prometheus_agent : sig ... end

module Prometheus_server : sig ... end

module Recipe : sig ... end

module Remote_agent : sig ... end

This module introduces the means to spawn and control agents on remote machine.

module Remote_procedure : sig ... end

module Remote_prometheus : sig ... end

module Services_cache : sig ... end

module Stage : sig ... end

A stage defines an ordered set of jobs to be executed on a set of agents known to the orchestrator.

module Template : sig ... end

module Tezos : sig ... end

module Uri : sig ... end

module Uri_resolver : sig ... end