Module Octogram.Agent_builtins

This module introduces the remote procedure allowing to manage the lifecycle of a remote agent.

val agent_should_continue : Agent_state.t -> bool

type start_http_server = {

directory : string;

http_port : string option;

}

type start_http_server_r = {

port : int;

}

type action =

| Create

| Extract

type 'uri tar = {

contents : string;

archive : 'uri;

action : action;

}

type Remote_procedure.t +=

| Quit : (unit, 'uri) Remote_procedure.t
(*
Quit can be used to terminate an agent gracefully.
*)

| Start_http_server : start_http_server -> (start_http_server_r, 'uri) Remote_procedure.t
(*
Start_http_server spawns a HTTP server on the agent, which can later be used to fetch files from it.

Each agent can spawn at most one HTTP server.
*)

| Tar : 'uri tar -> (unit, 'uri) Remote_procedure.t
(*
Tar requests an agent to Create or Extract an archive.

When Extracting, the archive can be on a remote agent. In such a case, the HTTP server of this agent will be used as a means to fetch the archive.
*)