Module Octogram.Execution_params

type mode =

| Sequential

| Concurrent

Should a sequence of tasks be executed in concurrently or sequentially?

val traverse : mode -> ('a -> unit Lwt.t) -> 'a Stdlib.Seq.t -> unit Lwt.t

traverse mode f seq iterates over each element of a lazy sequence seq either concurrently or sequentially, depending on the value of mode

val mode_encoding : mode Data_encoding.t