Module Octogram.Helpers

type case_maker = {

mk_case : 'a 'b. string -> 'a Data_encoding.t -> ('b -> 'a option) -> ('a -> 'b) -> 'b Data_encoding.case;

}

val make_mk_case : unit -> case_maker

val of_json_string : 'a Data_encoding.Encoding.t -> string -> 'a

val of_json_string_opt : 'a Data_encoding.Encoding.t -> string -> 'a option

val to_json_string : 'a Data_encoding.Encoding.t -> 'a -> string

val json_string_of_json : Ezjsonm.value -> string

val download : ?dir:string -> ?runner:Tezt__.Runner.t -> string -> string -> string Lwt.t

download url filename downloads the file at url, stores it in a temporary file named filename, and returns the complete path to the downloaded file.

The dir optional argument can be specified to the directory inside of which to save the file. If omitted, the directory is chosen by Tezt.

The runner optional argument can be specified to execute the command on a remote runner. If omitted, the command is executed locally.

val exec : ?can_fail:bool -> ?runner:Tezt__.Runner.t -> string -> string list -> unit Lwt.t

val mkdir_runnable : ?runner:Tezt__.Runner.t -> ?p:bool -> string -> (Tezt.Process.t, unit) Tezt_tezos.Runnable.t

val mkdir : ?runner:Tezt__.Runner.t -> ?p:bool -> string -> unit Lwt.t

val deploy_runnable : runner:Tezt.Runner.t -> ?r:bool -> string -> string -> (Tezt.Process.t, unit) Tezt_tezos.Runnable.t

val deploy : for_runner:Tezt.Runner.t -> ?r:bool -> string -> string -> unit Lwt.t

val when_ : bool -> unit Lwt.t -> unit Lwt.t

val change_tmp_directory_if_necessary : string option -> unit