Module Octogram.Orchestrator

val run_recipe : keep_alive:bool -> Recipe.t -> unit Lwt.t

run_recipe r unrolls the recipe r, that is it starts its agents, orchestrates its jobs, then clean-up.

Developers API

For developers interested in program Octogram scenarios though OCaml rather than the intended YAML recipe file.

val initialize_agents : octogram_binary:Remote_agent.octogram_binary -> state:Orchestrator_state.t -> Recipe.agent list -> unit Lwt.t

val terminate_agents : state:Orchestrator_state.t -> unit Lwt.t

val run_job : state:Orchestrator_state.t -> agent:Remote_agent.t -> re:Jingoo.Jg_types.tvalue -> string Job.t -> unit Lwt.t

val run_stage : state:Orchestrator_state.t -> Stage.t -> unit Lwt.t