Module type Remote_procedure.S

The module type to implement in order to be able to make agents capable of executing a new kind of remote procedure

val name : string

The name of the remote procedure, as referred to in the experience script when defining a new job.

type 'a t

The request sent by the orchestrator to the agent responsible to execute it.

type r

The result returned by the agent to the orchestrator.

val of_remote_procedure : ('a, 'uri) remote_procedure -> 'uri t option

val to_remote_procedure : 'uri t -> (r, 'uri) remote_procedure

val unify : ('a, 'uri) remote_procedure -> ('a, r) eq

val encoding : 'uri Data_encoding.t -> 'uri t Data_encoding.t

val r_encoding : r Data_encoding.t

val tvalue_of_r : r -> Jingoo.Jg_types.tvalue

val expand : self:Agent_name.t -> run:(string -> string) -> string t -> Uri.global_uri t

val resolve : self:Agent_name.t -> Uri_resolver.t -> Uri.global_uri t -> Uri.agent_uri t

val run : Agent_state.t -> Uri.agent_uri t -> r Lwt.t

val on_completion : on_new_service: (string -> Services_cache.node_kind -> Services_cache.service_kind -> int -> unit) -> on_new_metrics_source:(string -> Services_cache.node_kind -> int -> unit) -> r -> unit