Module Octogram.Stage

A stage defines an ordered set of jobs to be executed on a set of agents known to the orchestrator.

type t = {

name : string;

with_agents : string list;
(*
A list of regular expression used to select the subset of agents expected to run the jobs.
*)

run_agents : Execution_params.mode;
(*
Decide if the jobs of the stage has to be executed by each agent sequentially (there is at most one agent running jobs at a time) or concurrently (all agents start executing their job at the same time).
*)

run_jobs : Execution_params.mode;
(*
Decide if the job constituting a stage has to be run concurrently or sequentially by a given agent.
*)

ask_confirmation : bool;
(*
Decide whether or not the user has to request the start of the execution.
*)

jobs : string Job.t list;

}

val encoding : t Data_encoding.t